rename stats to report
