+++ picopt/walk/init.py
@@ -10,7 +10,7 @@ from picopt import PROGRAM_NAME
 from picopt.config.consts import TIMESTAMPS_CONFIG_KEYS
 from picopt.exceptions import PicoptError
 from picopt.old_timestamps import OldTimestamps
-from picopt.stats import Totals
+from picopt.report import Totals
 from picopt.walk.skip import Messenger, WalkSkipper
 
 if TYPE_CHECKING:
